text1 renamed to text again
