+++ components/text/index.js
